requirements.txt
pytest==7.2.0
allure-pytest==2.11.1
Appium-Python-Client==2.7.1
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
selene==2.0.0b14
pytest-xdist==2.5.0
